Rename: category controller 폴더 이동
