+++ src/api/category/category.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 
 import { CategoryService } from './service/category.service';
-import { CategoryController } from './category.controller';
+import { CategoryController } from './controller/category.controller';
 import { Category } from '../../entity/category.entity';
 
 @Module({
